requirements.txt
notebook
langchain
groq
ollama
langchain-google-genai
langchain-groq
google-genai
transformers
huggingface_hub[hf_xet]
python-dotenv
requests
pillow
ipykernel
openai
chainlit